% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confInts.R
\name{confInts}
\alias{confInts}
\title{confInts}
\usage{
confInts(v)
}
\arguments{
\item{v}{A vector}
}
\value{
Returns a confidence interval. For non-root nodes
}
\description{
Estimation helper function:
 Method that takes samples and generates confidence intervals for
 nodes other than the root. Assume raw data (not log), with normal distributed
 log data for confidence interval construction
}
\examples{
\donttest{
 data(treeData1)
 tree <- makeTree(treeData1)
 message("note - longer run time example")
 Zhats <- wmmTree(tree, sample_length = 10)
 confInts(Zhats$estimates)
}
}
