% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countTree.R
\name{countTree}
\alias{countTree}
\title{countTree}
\usage{
countTree(tree)
}
\arguments{
\item{tree}{A makeTree object}
}
\value{
Returns a tree plot
}
\description{
Visualize post-wmmTree tree with root estimate and marginal counts
 Also displays average of probability samples on each branch
}
\examples{
\donttest{
  message("note - longer run time example")
  data(treeData1)
  tree <- makeTree(treeData1)
  Zhats <- wmmTree(tree, sample_length = 3)
  countTree(tree)
 }
}
