% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawTree.R
\name{drawTree}
\alias{drawTree}
\title{drawTree}
\usage{
drawTree(tree, probs = TRUE, desc = TRUE)
}
\arguments{
\item{tree}{A makeTree object}

\item{probs}{A logical with default TRUE to specify whether to display
probabilities on branches}

\item{desc}{A logical with default TRUE to specify whether to display
node descriptions}
}
\value{
Returns a descriptive tree plot
}
\description{
Visualize tree with descriptions and probabilities; can be used
 pre-WMM analysis
}
\examples{
data(treeData1)
tree <- makeTree(treeData1)
drawTree(tree)
}
