% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estTree.R
\name{estTree}
\alias{estTree}
\title{estTree}
\usage{
estTree(tree)
}
\arguments{
\item{tree}{A makeTree object}
}
\value{
Returns a tree plot
}
\description{
Visualize post-wmmTree tree with root estimate given by each branch,
 and weighted sum at the root (for post-analysis).
 Also displays average of probability samples on each branch.
}
\examples{
\donttest{
  message("note - longer run time example")
  data(treeData1)
  tree <- makeTree(treeData1)
  Zhats <- wmmTree(tree, sample_length = 3)
  estTree(tree)
 }
}
