% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ko.weights.R
\name{ko.weights}
\alias{ko.weights}
\title{ko.weights}
\usage{
ko.weights(tree)
}
\arguments{
\item{tree}{A makeTree object}
}
\value{
Returns vector of variance-minimizing weights on informative paths
}
\description{
Estimation helper function:
 Calculating variance minimizing weights.  Only assigns weights to informative paths
}
\examples{
\dontrun{
 ## example admissible dataset
 data(treeData1)
 tree <- makeTree(treeData1)
 ko.weights(tree)
}
}
