% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logEstimates.R
\name{logEstimates}
\alias{logEstimates}
\title{logEstimates}
\usage{
logEstimates(tree)
}
\arguments{
\item{tree}{A makeTree object}
}
\value{
Returns a vector of mean log estimate values of the root population size
 from each informative path
}
\description{
Estimation helper function:
 Creates a vector of mean estimate values given by each informative path
}
\examples{
\dontrun{
  ## example admissible dataset
 data(treeData1)
 tree <- makeTree(treeData1)
 Zhats <- wmmTree(tree, sample_length = 10)
 logEstimates(tree)
}
}
