% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmEstimate.R
\name{mmEstimate}
\alias{mmEstimate}
\title{mmEstimate}
\usage{
mmEstimate(o)
}
\arguments{
\item{o}{A node from a makeTree object}
}
\value{
Returns node with modified attributes
}
\description{
Helper function:
 Performs the multiplier method from a single terminal node (o) and
 returns the root estimate given that path, and the probabilities
 of each branch on that path.
}
\examples{
\donttest{
 data(treeData1)
 tree <- makeTree(treeData1)
 mmEstimate(tree$A)
 mmEstimate(tree$B)
 tree$A$targetEst
 tree$B$targetEst
}
}
