% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Nhats.R
\name{Nhats}
\alias{Nhats}
\title{Nhats}
\usage{
Nhats(tree)
}
\arguments{
\item{tree}{A makeTree object}
}
\value{
Returns a vector of root population size estimates from each sample
 run of the wmmTree function
}
\description{
Estimation helper function:
 Returns Nhat for each sample from the WMM (rather than the aggregate
 value given by the average, this calculates weights and applies the
 weighted sum to each of the samples)
}
\examples{
\donttest{
 data(treeData1)
 tree <- makeTree(treeData1)
 Zhats <- wmmTree(tree, sample_length = 3)
 Nhats(tree)
}
}
