% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/root.confInt.R
\name{root.confInt}
\alias{root.confInt}
\title{root.confInt}
\usage{
root.confInt(tree, int.type = "quantiles")
}
\arguments{
\item{tree}{A makeTree object}

\item{int.type}{A string specifying interval type, passed from the wmmTree
function.}
}
\value{
Returns a confidence interval for the root population size estimate
 in un-logged form.
}
\description{
Estimation helper function:
 Method for generating confidence interval for root node. Assumes unlogged
 input data and normally distributed logged data.  Completes conversion
 internally.
}
\examples{
\donttest{
 message("note - longer run time example")
 data(treeData1)
 tree <- makeTree(treeData1)
 Zhats <- wmmTree(tree, sample_length = 3)
 root.confInt(tree)
}
}
