% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleBeta.R
\name{sampleBeta}
\alias{sampleBeta}
\title{sampleBeta}
\usage{
sampleBeta(x, n, pop, node)
}
\arguments{
\item{x}{An integer; typical use case is survey numerator}

\item{n}{An integer; typical use case is survey sample size}

\item{pop}{A logical value which takes TRUE if sample size is population size}

\item{node}{A node from a makeTree object; carried forever from}
}
\value{
Returns sample from Beta distribution with parameters dependent on x, n
}
\description{
Helper function:
 Method for sampling from a Beta distribution given the survey estimates
}
\examples{
\donttest{
 data(treeData1)
 tree <- makeTree(treeData1)
 Zhats <- wmmTree(tree, sample_length = 3)
 sampleBeta(10, 55, pop = FALSE, tree$A)
}
}
