% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss.confInts.R
\name{ss.confInts}
\alias{ss.confInts}
\title{ss.confInts}
\usage{
ss.confInts(o, digits = 3)
}
\arguments{
\item{o}{A node of a makeTree object}

\item{digits}{The number of significant digits to report}
}
\value{
A confidence interval for nodes of the tree that only use a single
  source of sibling data.
}
\description{
Estimation helper function:
 Method that takes samples and generates confidence intervals for
 nodes in single source sibling tree (single.source = TRUE)
}
\examples{
\donttest{
 data(treeData1)
 tree <- makeTree(treeData1)
 Zhats <- wmmTree(tree, sample_length = 3)
 ss.confInts(tree$B)
}
}
