% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssEstimate.R
\name{ssEstimate}
\alias{ssEstimate}
\title{ssEstimate}
\usage{
ssEstimate(o)
}
\arguments{
\item{o}{A node from a makeTree object}
}
\value{
Returns node with modified attributes
}
\description{
Helper function:
 Performs the closed form calculation of variances and means based on a
 "single-source sibling" tree.  Engages when single.source = TRUE
 in wmmTree function.  See documentation for further details.
}
\examples{
\donttest{
 data(treeData1)
 tree <- makeTree(treeData1)
 Zhats <- wmmTree(tree, sample_length = 3)
 ssEstimate(tree$B)
}
}
