% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeData1.R
\docType{data}
\name{treeData1}
\alias{treeData1}
\title{Simple Tree Data}
\format{
An object of class \code{"data.frame"}
\describe{
 \item{from}{A node label and started point of directed edge (parent node)}
 \item{to}{A node label and endpoint of directed edge (child node)}
 \item{Estimate}{A numerical value assumed to be survey count belonging to 'to' node (integer)}
 \item{Total}{A numerical value assumed to be survey sample size (integer)}
 \item{Count}{A numerical value for marginal count if leaf node (integer)}
 \item{Population}{A boolean value for if survey size is entire population (logical)}
 \item{Description}{A string describing 'to' node (string)}
}
}
\usage{
data(treeData1)
}
\description{
Small, artificially generated toy data set to demonstrate package functionality
}
\examples{
data(treeData1)
head(treeData1)

}
\references{
This data set was artificially created for the AutoWMM package.
}
\keyword{datasets}
