% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeData2.R
\docType{data}
\name{treeData2}
\alias{treeData2}
\title{Larger Tree Data}
\format{
An object of class \code{"data.frame"}
\describe{
 \item{from}{A node label and started point of directed edge (parent node)}
 \item{to}{A node label and endpoint of directed edge (child node)}
 \item{Estimate}{A numerical value assumed to be survey count belonging to 'to' node (integer)}
 \item{Total}{A numerical value assumed to be survey sample size (integer)}
 \item{Count}{A numerical value for marginal count if leaf node (integer)}
}
}
\usage{
data(treeData2)
}
\description{
Larger artificially generated toy data set to demonstrate package functionality
}
\examples{
data(treeData2)
head(treeData2)

}
\references{
This data set was artificially created for the AutoWMM package.
}
\keyword{datasets}
